Calculator functions reorganized.
