--- colorize.rb
@@ -1,19 +1,0 @@
-module CmdCalculator
-  class Colorize
-    def colorize(text, color_code)
-      "\e[#{color_code}m#{text}\e[0m"
-    end
-
-    def red(text)
-      colorize(text, 31)
-    end
-
-    def green(text)
-      colorize(text, 32)
-    end
-
-    def blue(text)
-      colorize(text, 34)
-    end
-  end
-end
